src/app/core/agents/tools.py
"""Tools available to agents in the multi-agent RAG system."""

from langchain_core.tools import tool

from ..retrieval.vector_store import retrieve
from ..retrieval.serialization import serialize_chunks_with_ids


@tool(response_format="content_and_artifact")
def retrieval_tool(query: str):
    """Search the vector database for relevant document chunks.

    This tool retrieves the top 4 most relevant chunks from the Pinecone
    vector store based on the query. The chunks are formatted with stable
    IDs [C1], [C2] for citation.

    Args:
        query: The search query string to find relevant document chunks.

    Returns:
        Tuple of (serialized_content, artifact) where:
        - serialized_content: Formatted string with chunks and [ID] tags.
        - artifact: Dictionary containing 'docs' (raw objects) and 'citations' (metadata map).
    """
    # Retrieve documents from vector store
    docs = retrieve(query, k=4)

    # Serialize chunks into formatted string with stable IDs
    context, citation_map = serialize_chunks_with_ids(docs)

    # Return tuple: (serialized content, artifact dictionary)
    return context, {
        "docs": docs,
        "citations": citation_map
    }
