pyproject.toml
[project]
name = "al-bio-question-answer-resource-book"
version = "0.1.0"
description = "A/L Biology Question Answer RAG System for Sri Lankan students"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.124.0",
    "langchain>=1.1.2",
    "langchain-community>=0.3.0",
    "langchain-openai>=1.1.0",
    "langchain-pinecone>=0.2.13",
    "langchain-text-splitters>=1.0.0",
    "langgraph>=1.0.4",
    "pinecone>=3.0.0",
    "pydantic-settings>=2.0.0",
    "pypdf>=6.4.1",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.20",
    "uvicorn>=0.38.0",
]
